Add grunt-nunjucks template pre-compilation
