+++ package.json
@@ -20,7 +20,8 @@
     "underscore": "~1.5.2",
     "grunt-contrib-watch": "~0.5.3",
     "connect-livereload": "~0.3.2",
-    "http-rewrite-middleware": "~0.1.3"
+    "http-rewrite-middleware": "~0.1.3",
+    "grunt-nunjucks": "~0.1.3"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
