+++ public/js/views/login.js
@@ -3,7 +3,7 @@ define(['views/base', 'log'], function(BaseView, log){
   var LoginView = BaseView.extend({
     render: function(){
       console.log('rendering view');
-      this.setTitle('Sign In');
+      this.setTitle(this.gettext('Sign In'));
       this.renderTemplate('login.html', {heading: 'Sign In'});
       return this;
     }
